Getting Started - Starting client and server + Notes on design
HELP.md file contains data that will be useful to complete the README which is lacking some critical information to run the app.

Startup now check for AllowOpenETSync to avoid a crash when OpenETClient is not setup in the environment.

Added DB_CONNECTION_STRING to be set in order to connect to the SQL server from the container.